+++ aibRDSUnifiedDesktop.ps1
@@ -46,7 +46,9 @@ Start-AzImageBuilderTemplate -ResourceGroupName $imageResourceGroup -Name $image
 $gallery = Get-AzGallery -Name $galleryName
 $versions = Get-AzGalleryImageVersion -ResourceGroupName $gallery.ResourceGroupName -GalleryName $gallery.Name -GalleryImageDefinitionName $imageTemplateName
 $oldestVersion = $versions | Sort-Object -Property Name | Select-Object -First 1
-"Found oldest version $($oldestVersion.Name)...Deleting..."
-$oldestVersion | Remove-AzGalleryImageVersion -Force
+if ($versions.count -gt 3) {
+    "Found oldest version $($oldestVersion.Name)...Deleting..."
+    $oldestVersion | Remove-AzGalleryImageVersion -Force
 
+}
